requirements.txt
fastapi[all]
pydantic
pytest
pytest-cov
pymongo[srv]
python-dotenv
python-jose
bcrypt
uvicorn
flake8
